chore: cleanup db connectors setup
- refactor the db connectors to remove a lot of duplicate code and make it easier to add DB connectors in the future
- add documentation
- remove dead code
- add generic error messages
