+++ csml_manager/src/db_connectors/http/interactions.rs
