+++ csml_manager/src/db_connectors/http/state.rs
