+++ csml_manager/src/error_messages.rs
@@ -1,0 +1,1 @@
+pub const ERROR_DB_SETUP: &'static str = "Database connector is not setup correctly";
