+++ csml_manager/src/lib.rs
@@ -4,8 +4,9 @@ pub use csml_interpreter::data::{
 };
 use serde_json::json;
 
-mod db_interactions;
-#[cfg(any(feature = "mongo"))]
+mod error_messages;
+mod db_connectors;
+
 mod encrypt;
 mod init;
 mod interpreter_actions;
@@ -13,7 +14,7 @@ mod send;
 mod tools;
 
 use data::*;
-use db_interactions::{conversation::*, init_db, messages::*, state::*, Conversation};
+use db_connectors::{conversation::*, init_db, messages::*, state::*, Conversation};
 use init::*;
 use interpreter_actions::interpret_step;
 use tools::*;
